requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.14.0
pinecone-client==5.0.1
python-dotenv==1.0.0
pydantic==2.6.0
httpx==0.25.2
google-generativeai==0.8.3